Use long for strtol()

+++ mta/date.c
@@ -51,7 +51,7 @@ time_t smtp_getTime(const char *b, unsigned char * const tzp) {
 	offset += 3;
 	while (b[offset] == ' ') offset++;
 
-	int year = strtol(b + offset, &end, 10);
+	long year = strtol(b + offset, &end, 10);
 	if (year > 20 && year < 100) year += 2000;
 	if (year < 2020) return 0;
 
@@ -59,20 +59,20 @@ time_t smtp_getTime(const char *b, unsigned char * const tzp) {
 	offset = 1;
 	while (b[offset] == ' ') offset++;
 
-	const int hour = strtol(b + offset, &end, 10);
+	const long hour = strtol(b + offset, &end, 10);
 	if (*end != ':') return 0;
 	b = end;
 	offset = 1;
 	while (b[offset] == ' ') offset++;
 
-	const int min = strtol(b + offset, &end, 10);
+	const long min = strtol(b + offset, &end, 10);
 	if (*end != ':') return 0;
 	b = end;
 	offset = 1;
 
 	while (b[offset] == ' ') offset++;
 
-	const int sec = strtol(b + offset, &end, 10);
+	const long sec = strtol(b + offset, &end, 10);
 	if (*end != ' ') return 0;
 	b = end;
 	offset = 0;
